+++ include/Register/DelayRegisterWord.h
@@ -13,11 +13,11 @@ public:
         : RegisterWord ("DelayRegisterWord"), next_addr (0),
             addr_sub(0), val16(0) {}
 
-    DelayRegisterWord (uint64_t next_addr, uint16_t addr_sub, uint16_t val16)
+    DelayRegisterWord (uint32_t next_addr, uint64_t addr_sub, uint16_t val16)
         : RegisterWord ("DelayRegisterWord"), next_addr (next_addr),
             addr_sub (addr_sub), val16 (val16) {}
     
-    uint64_t next_addr;
-    uint16_t addr_sub;
+    uint32_t next_addr;
+    uint64_t addr_sub;
     uint16_t val16;
 };
